# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Tracker API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - owner: []
paths:
...
  /goals:
    {
      'get':
        {
          'operationId': 'listGoals',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            { 'type': 'array', 'items': { '$ref': '#/components/schemas/Goal' } },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
        },
      'post':
        {
          'operationId': 'createGoal',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    { 'application/json': { 'schema': { '$ref': '#/components/schemas/Goal' } } },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'requestBody':
            {
              'required': true,
              'content':
                { 'application/json': { 'schema': { '$ref': '#/components/schemas/GoalInput' } } },
            },
        },
    }
  /conversations/{conversationId}/outcomes:
    {
      'get':
        {
          'operationId': 'listConversationOutcomes',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'array',
                              'items': { '$ref': '#/components/schemas/Outcome' },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'conversationId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
        },
      'post':
        {
          'operationId': 'recordConversationOutcome',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json': { 'schema': { '$ref': '#/components/schemas/Outcome' } },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'conversationId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
          'requestBody':
            {
              'required': true,
              'content':
                {
                  'application/json': { 'schema': { '$ref': '#/components/schemas/OutcomeInput' } },
                },
            },
        },
    }
  /outcomes/{outcomeId}/reverse:
    {
      'post':
        {
          'operationId': 'reverseOutcome',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'object',
                              'required': ['reversed'],
                              'properties': { 'reversed': { 'type': 'boolean' } },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'outcomeId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
  /reports:
    {
      'get':
        {
          'operationId': 'getAnalyticsReport',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        { 'schema': { '$ref': '#/components/schemas/AnalyticsReport' } },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'from',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string', 'format': 'date-time' },
              },
              {
                'name': 'to',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string', 'format': 'date-time' },
              },
              {
                'name': 'mailboxId',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string' },
              },
              {
                'name': 'goalId',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
  /reports/export:
    {
      'get':
        {
          'operationId': 'exportAnalyticsReport',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content': { 'text/csv': { 'schema': { 'type': 'string' } } },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'from',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string', 'format': 'date-time' },
              },
              {
                'name': 'to',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string', 'format': 'date-time' },
              },
              {
                'name': 'mailboxId',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string' },
              },
              {
                'name': 'goalId',
                'in': 'query',
                'required': false,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
  /reminders:
    {
      'get':
        {
          'operationId': 'listReminders',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'array',
                              'items': { '$ref': '#/components/schemas/Reminder' },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
        },
    }
  /conversations/{conversationId}/reminders:
    {
      'post':
        {
          'operationId': 'createReminder',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'object',
                              'required': ['created'],
                              'properties': { 'created': { 'type': 'boolean' } },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'conversationId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
          'requestBody':
            {
              'required': true,
              'content':
                {
                  'application/json':
                    {
                      'schema':
                        {
                          'type': 'object',
                          'required': ['dueAt'],
                          'properties': { 'dueAt': { 'type': 'string', 'format': 'date-time' } },
                        },
                    },
                },
            },
        },
    }
  /reminders/{reminderId}/done:
    {
      'post':
        {
          'operationId': 'completeReminder',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'object',
                              'required': ['done'],
                              'properties': { 'done': { 'type': 'boolean' } },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'reminderId',
                'in': 'path',
                'required': true,
                'schema': { 'type': 'string' },
              },
            ],
        },
    }
...
  /webhooks/conversions:
    {
      'post':
        {
          'operationId': 'recordConversionWebhook',
          'responses':
            {
              '200':
                {
                  'description': 'Successful operation',
                  'content':
                    {
                      'application/json':
                        {
                          'schema':
                            {
                              'type': 'object',
                              'required': ['recorded', 'attributed', 'id'],
                              'properties':
                                {
                                  'recorded': { 'type': 'boolean' },
                                  'attributed': { 'type': 'boolean' },
                                  'id': { 'type': 'string' },
                                },
                            },
                        },
                    },
                },
              '400': { 'description': 'Invalid input' },
              '401': { 'description': 'Authentication required' },
              '404': { 'description': 'Resource not found' },
              '429': { 'description': 'Rate limit reached' },
              '500': { 'description': 'Internal error' },
            },
          'parameters':
            [
              {
                'name': 'X-Webhook-Timestamp',
                'in': 'header',
                'required': false,
                'schema': { 'type': 'string' },
              },
              {
                'name': 'X-Webhook-Signature',
                'in': 'header',
                'required': false,
                'schema': { 'type': 'string' },
              },
            ],
          'requestBody':
            {
              'required': true,
              'content':
                {
                  'application/json': { 'schema': { '$ref': '#/components/schemas/OutcomeInput' } },
                },
            },
          'security': [],
          'description': 'HMAC-SHA256 of timestamp + dot + exact body using CONVERSION_WEBHOOK_SECRET. Timestamp tolerance five minutes. externalId required and idempotent. Unmatched attribution is retained without a delivery.',
        },
    }
components:
  securitySchemes:
    owner: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Goal:
      {
        'type': 'object',
        'required': ['id', 'name', 'windowDays'],
        'properties':
          {
            'id': { 'type': 'string' },
            'name': { 'type': 'string' },
            'windowDays': { 'type': 'integer' },
          },
      }
    GoalInput:
      {
        'type': 'object',
        'required': ['name', 'windowDays'],
        'properties':
          {
            'name': { 'type': 'string', 'minLength': 1, 'maxLength': 100 },
            'windowDays': { 'type': 'integer', 'minimum': 1, 'maximum': 365 },
          },
      }
    OutcomeInput:
      {
        'type': 'object',
        'required': ['goalId', 'occurredAt'],
        'properties':
          {
            'goalId': { 'type': 'string', 'format': 'uuid' },
            'occurredAt': { 'type': 'string', 'format': 'date-time' },
            'deliveryId': { 'type': 'string' },
            'attributionRef': { 'type': 'string' },
            'externalId': { 'type': 'string', 'maxLength': 128 },
            'value': { 'type': 'string', 'pattern': "^(0|[1-9][0-9]{0,12})(\\.[0-9]{1,2})?$" },
            'currency': { 'type': 'string', 'pattern': '^[A-Z]{3}$' },
          },
      }
    Outcome:
      {
        'type': 'object',
        'required':
          ['id', 'goalId', 'deliveryId', 'source', 'occurredAt', 'reversedAt', 'value', 'currency'],
        'properties':
          {
            'id': { 'type': 'string' },
            'goalId': { 'type': 'string' },
            'deliveryId': { 'type': 'string', 'nullable': true },
            'source': { 'type': 'string' },
            'occurredAt': { 'type': 'string', 'format': 'date-time' },
            'reversedAt': { 'type': 'string', 'format': 'date-time', 'nullable': true },
            'value': { 'type': 'string', 'nullable': true },
            'currency': { 'type': 'string', 'nullable': true },
          },
      }
    AnalyticsReport:
      {
        'type': 'object',
        'required':
          [
            'sent',
            'imageActive',
            'linked',
            'clicked',
            'replied',
            'bounceReported',
            'converted',
            'prepared',
            'unknown',
            'from',
            'to',
            'policy',
          ],
        'properties':
          {
            'sent': { 'type': 'integer' },
            'imageActive': { 'type': 'integer' },
            'linked': { 'type': 'integer' },
            'clicked': { 'type': 'integer' },
            'replied': { 'type': 'integer' },
            'bounceReported': { 'type': 'integer' },
            'converted': { 'type': 'integer' },
            'prepared': { 'type': 'integer' },
            'unknown': { 'type': 'integer' },
            'from': { 'type': 'string', 'format': 'date-time' },
            'to': { 'type': 'string', 'format': 'date-time' },
            'policy': { 'type': 'string' },
          },
      }
...
    Reminder:
      {
        'type': 'object',
        'required': ['id', 'conversationId', 'subject', 'dueAt'],
        'properties':
          {
            'id': { 'type': 'string' },
            'conversationId': { 'type': 'string' },
            'subject': { 'type': 'string' },
            'dueAt': { 'type': 'string', 'format': 'date-time' },
          },
      }
...